requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docopt==0.6.2
google-auth==2.25.2
helm-client==0.3.0
idna==3.6
kubernetes==28.1.0
oauthlib==3.2.2
packaging==23.2
pick==2.2.0
pipreqs==0.4.13
psycopg
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
websocket-client==1.7.0
yarg==0.1.9
